requirements.txt
setuptools>=54.2.0
phonenumbers>=8.12.15
email_validator
typing>=3.7.4.3
Hypercorn>=0.11.2

starlette>=0.13.6
httpx>=0.18.1
pydantic~=1.8.2
validators~=0.18.1
ipaddress~=1.0.23
loguru~=0.5.3
fastapi~=0.65.2
python-dotenv~=0.17.1
pytest~=6.2.2
CherryPy~=18.6.1
response-report~=1.1.0
